Move Profile Panel to Profile Component Directory
